// Kraftvaerk.Umbraco.Headless.BlockPreview.Extension/src/blockpreview-api/models/HeadlessPreviewToggleModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type HeadlessPreviewToggleModel = {
    id: string;
    enabled: boolean;
    enabledNested: boolean;
    enabledList: boolean;
    enabledGrid: boolean;
    enabledRTE: boolean;
    advancedSettings?: string | null;
    alias?: string | null;
};